Refactor: account router (#1063)
