+++ backend/src/database/queries.ts
@@ -425,245 +425,6 @@ export const queryActiveAccountsList = async () => {
     .execute();
 };
 
-export const queryIndexedAccount = async (accountId: string) => {
-  return indexerDatabase
-    .selectFrom("accounts")
-    .select("account_id")
-    .where("account_id", "=", accountId)
-    .limit(1)
-    .executeTakeFirst();
-};
-
-export const queryAccountsList = async (
-  limit: number = 15,
-  cursor?: number
-) => {
-  let selection = indexerDatabase
-    .selectFrom("accounts")
-    .select(["account_id", "id as account_index"])
-    .leftJoin("receipts", (join) =>
-      join.onRef("receipt_id", "=", "created_by_receipt_id")
-    );
-  if (cursor !== undefined) {
-    selection = selection.where("id", "<", cursor.toString());
-  }
-  return selection.orderBy("account_index", "desc").limit(limit).execute();
-};
-
-export const queryAccountFungibleTokenContractIds = async (
-  accountId: string
-): Promise<string[]> => {
-  const selection = await indexerDatabase
-    .selectFrom("assets__fungible_token_events")
-    .select("emitted_by_contract_account_id as contractId")
-    .distinctOn("emitted_by_contract_account_id")
-    .where("token_new_owner_account_id", "=", accountId)
-    .orWhere("token_old_owner_account_id", "=", accountId)
-    .orderBy("emitted_by_contract_account_id", "desc")
-    .execute();
-  return selection.map((row) => row.contractId);
-};
-
-export const queryAccountFungibleTokenHistory = async (
-  accountId: string,
-  tokenAuthorAccountId: string
-) => {
-  const selection = await indexerDatabase
-    .selectFrom("assets__fungible_token_events")
-    .innerJoin("receipts", (qb) =>
-      qb.onRef("emitted_for_receipt_id", "=", "receipts.receipt_id")
-    )
-    .innerJoin("blocks", (qb) =>
-      qb.onRef("blocks.block_hash", "=", "receipts.included_in_block_hash")
-    )
-    .select([
-      "amount",
-      "token_old_owner_account_id as prevAccountId",
-      "token_new_owner_account_id as nextAccountId",
-      "emitted_for_receipt_id as receiptId",
-      (eb) => div(eb, "emitted_at_block_timestamp", 1000 * 1000, "timestamp"),
-      "originated_from_transaction_hash as transactionHash",
-      "block_height as blockHeight",
-    ])
-    .where("emitted_by_contract_account_id", "=", tokenAuthorAccountId)
-    .where((wi) =>
-      wi
-        .where("token_new_owner_account_id", "=", accountId)
-        .orWhere("token_old_owner_account_id", "=", accountId)
-    )
-    .orderBy("emitted_at_block_timestamp", "desc")
-    .orderBy("emitted_in_shard_id", "desc")
-    .orderBy("emitted_index_of_event_entry_in_shard", "desc")
-    // Pagination to be introduced soon..
-    .limit(200)
-    .execute();
-  return selection;
-};
-
-export const queryOutcomeTransactionsCountFromAnalytics = async (
-  accountId: string
-) => {
-  const selection = await analyticsDatabase
-    .selectFrom("daily_outgoing_transactions_per_account_count")
-    .select([
-      (eb) =>
-        sum(eb, "outgoing_transactions_count").as("out_transactions_count"),
-      (eb) => max(eb, "collected_for_day").as("last_day_collected"),
-    ])
-    .where("account_id", "=", accountId)
-    .executeTakeFirstOrThrow();
-  return {
-    out_transactions_count: selection.out_transactions_count
-      ? parseInt(selection.out_transactions_count)
-      : 0,
-    last_day_collected_timestamp: selection.last_day_collected
-      ? millisecondsToNanoseconds(
-          selection.last_day_collected.getTime() + DAY
-        ).toString()
-      : undefined,
-  };
-};
-
-export const queryOutcomeTransactionsCountFromIndexerForLastDay = async (
-  accountId: string,
-  lastDayCollectedTimestamp?: string
-) => {
-  // since analytics are collected for the previous day,
-  // then 'lastDayCollectedTimestamp' may be 'null' for just created accounts so
-  // we must put 'lastDayCollectedTimestamp' as below to dislay correct value
-  const timestamp =
-    lastDayCollectedTimestamp ||
-    millisecondsToNanoseconds(new Date().getTime() - DAY).toString();
-  const selection = await indexerDatabase
-    .selectFrom("transactions")
-    .select((eb) => count(eb, "transaction_hash").as("out_transactions_count"))
-    .where("signer_account_id", "=", accountId)
-    .where("block_timestamp", ">=", timestamp)
-    .executeTakeFirstOrThrow();
-  return parseInt(selection.out_transactions_count);
-};
-
-export const queryIncomeTransactionsCountFromAnalytics = async (
-  accountId: string
-) => {
-  const selection = await analyticsDatabase
-    .selectFrom("daily_ingoing_transactions_per_account_count")
-    .select([
-      (eb) => sum(eb, "ingoing_transactions_count").as("in_transactions_count"),
-      (eb) => max(eb, "collected_for_day").as("last_day_collected"),
-    ])
-    .where("account_id", "=", accountId)
-    .executeTakeFirstOrThrow();
-  return {
-    in_transactions_count: selection.in_transactions_count
-      ? parseInt(selection.in_transactions_count)
-      : 0,
-    last_day_collected_timestamp: selection.last_day_collected
-      ? millisecondsToNanoseconds(
-          selection.last_day_collected.getTime() + DAY
-        ).toString()
-      : undefined,
-  };
-};
-
-export const queryIncomeTransactionsCountFromIndexerForLastDay = async (
-  accountId: string,
-  lastDayCollectedTimestamp?: string
-) => {
-  // since analytics are collected for the previous day,
-  // then 'lastDayCollectedTimestamp' may be 'null' for just created accounts so
-  // we must put 'lastDayCollectedTimestamp' as below to dislay correct value
-  const timestamp =
-    lastDayCollectedTimestamp ||
-    millisecondsToNanoseconds(new Date().getTime() - DAY).toString();
-  const selection = await indexerDatabase
-    .selectFrom("transactions")
-    // TODO: Research if we can get rid of distinct without performance degradation
-    .select(
-      sql<string>`count(distinct transactions.transaction_hash)`.as(
-        "in_transactions_count"
-      )
-    )
-    .leftJoin("receipts", (jb) =>
-      jb
-        .onRef("originated_from_transaction_hash", "=", "transaction_hash")
-        .on("block_timestamp", ">=", timestamp)
-    )
-    .where("included_in_block_timestamp", ">=", timestamp)
-    .where("signer_account_id", "!=", accountId)
-    .where("receipts.receiver_account_id", "=", accountId)
-    .executeTakeFirst();
-  if (!selection) {
-    return 0;
-  }
-  return parseInt(selection.in_transactions_count);
-};
-
-export const queryAccountInfo = async (accountId: string) => {
-  const selection = await indexerDatabase
-    .selectFrom((eb) =>
-      eb
-        .selectFrom("accounts")
-        .select([
-          "account_id",
-          "created_by_receipt_id",
-          "deleted_by_receipt_id",
-        ])
-        .where("account_id", "=", accountId)
-        .as("inneraccounts")
-    )
-    .leftJoin("receipts as creation_receipt", (jb) =>
-      jb.onRef("creation_receipt.receipt_id", "=", "created_by_receipt_id")
-    )
-    .leftJoin("receipts as deletion_receipt", (jb) =>
-      jb.onRef("deletion_receipt.receipt_id", "=", "deleted_by_receipt_id")
-    )
-    .where("account_id", "=", accountId)
-    .select([
-      "account_id",
-      (eb) =>
-        div(
-          eb,
-          "creation_receipt.included_in_block_timestamp",
-          1000 * 1000,
-          "created_at_block_timestamp"
-        ),
-      "creation_receipt.originated_from_transaction_hash as created_by_transaction_hash",
-      (eb) =>
-        div(
-          eb,
-          "deletion_receipt.included_in_block_timestamp",
-          1000 * 1000,
-          "deleted_at_block_timestamp"
-        ),
-      "deletion_receipt.originated_from_transaction_hash as deleted_by_transaction_hash",
-    ])
-    .executeTakeFirst();
-  if (!selection) {
-    return;
-  }
-  return {
-    // TODO: Discover how to get rid of non-null type assertion
-    accountId: selection.account_id!,
-    created:
-      selection.created_at_block_timestamp &&
-      selection.created_by_transaction_hash
-        ? {
-            timestamp: selection.created_at_block_timestamp,
-            hash: selection.created_by_transaction_hash,
-          }
-        : undefined,
-    deleted:
-      selection.deleted_at_block_timestamp &&
-      selection.deleted_by_transaction_hash
-        ? {
-            timestamp: selection.deleted_at_block_timestamp,
-            hash: selection.deleted_by_transaction_hash,
-          }
-        : undefined,
-  };
-};
-
 // contracts
 export const queryNewContractsCountAggregatedByDate = async () => {
   return analyticsDatabase
