+++ common/src/index.ts
@@ -3,3 +3,4 @@ export * from "./utils/promise";
 export * from "./utils/git";
 export * from "./utils/queries";
 export * from "./utils/near";
+export * from "./utils/utils";
