+++ common/src/utils/utils.ts
@@ -1,0 +1,3 @@
+export const notNullGuard = <T>(arg: T): arg is Exclude<T, null> => {
+  return arg !== null;
+};
