+++ frontend/src/components/accounts/AccountRow.tsx
@@ -53,7 +53,10 @@ export interface Props {
 
 const AccountRow: React.FC<Props> = React.memo(({ accountId }) => {
   const { t } = useTranslation();
-  const { data: accountInfo } = trpc.useQuery(["account-info", [accountId]]);
+  const { data: accountInfo } = trpc.useQuery([
+    "account.byIdOld",
+    { id: accountId },
+  ]);
 
   return (
     <Link href={`/accounts/${accountId}`} passHref>
