+++ frontend/src/pages/accounts/[id].tsx
@@ -31,7 +31,7 @@ const AccountDetail: NextPage = React.memo(() => {
   useAnalyticsTrackOnMount("Explorer View Individual Account", {
     accountId,
   });
-  const accountQuery = trpc.useQuery(["account-info", [accountId]]);
+  const accountQuery = trpc.useQuery(["account.byIdOld", { id: accountId }]);
   const query = trpc.useInfiniteQuery(
     [
       "transactions-list-by-account-id",
